feat: 기존 URI 로깅에서 Latency 정보 기능 추가
