+++ server/src/main/java/com/project/yozmcafe/config/HibernateConfiguration.java
@@ -1,0 +1,23 @@
+package com.project.yozmcafe.config;
+
+import com.project.yozmcafe.controller.filter.QueryInspector;
+import org.hibernate.cfg.AvailableSettings;
+import org.springframework.boot.autoconfigure.orm.jpa.HibernatePropertiesCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class HibernateConfiguration {
+
+    private final QueryInspector queryInspector;
+
+    public HibernateConfiguration(final QueryInspector queryInspector) {
+        this.queryInspector = queryInspector;
+    }
+
+    @Bean
+    public HibernatePropertiesCustomizer hibernatePropertiesCustomizer() {
+        return hibernateProperties ->
+                hibernateProperties.put(AvailableSettings.STATEMENT_INSPECTOR, queryInspector);
+    }
+}
